Install BackstopJS as a dependency
I refuse to go the global route. I want to install everything as a recorded
dependency. So I use the user/repo instead of the package name:

```shell
npm install --save-dev garris/BackstopJS
```

This downloads the package to `node_modules` and the `--save-dev` records
it in `package.json` so we can track the dependency in our project.

Because npm was used to install rather than bower, dependencies were
installed recursively.

Adds `node_modules` to `.gitignore` file now that we have a dependency.