Generate Reference Files
Adds script to package.json so we can run BackstopJS gulp tasks as `npm run
backstop <task>` but found that it only works from the BackstopJS
directory. Set the cwd with argument to gulp. But then found it couldn't
find the local html file with that path. So the path was changed to
`../../myCoolProject/index.html`.

Generates reference files like so:

```shell
npm run backstop reference
```

They are saved in a location that is ignored but they should be there at
this point in `node_modules/BackstopJS/bitmaps_reference`.

You can run the first test like so:

```shell
npm run backstop test
```

This will generate test images in `node_modules/BackstopJS/bitmaps_test`
and open a report in the browser.